Organize imports
Ok so the Jool 4 prototype is ready, and it seems that the
migration is going to address even more quirks than originally
intended. Device driver mode appears to be hands down a much
superior model and I'm at this point abandoning the idea of
maintaining Netfilter and iptables support.

Instead of creating a small macro maze for the sake of preserving
previous operation modes, I will actually start dropping the
Netfilter code.

This code is still far from stable.